pyproject.toml
[tool.poetry]
name = "salary-service"
version = "0.1.0"
description = "REST service for viewing salary information"
authors = ["Vitaliy Guselnikov <dinwin1408@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = ">=0.99.1"
uvicorn = ">=0.22.0"
psycopg2-binary = ">=2.9.6"
pydantic = ">=1.10.10"
Jinja2 = "^3.1.2"
httpx = "^0.24.1"
python-multipart = "^0.0.6"

[tool.poetry.dev-dependencies]
pytest = "^7.4.0"
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
